refactored changes
